+++ packages/web/src/components/SegmentedControl/SegmentedControlOption.tsx
@@ -23,17 +23,17 @@ export const SegmentedControlOption = (props: SegmentedControlOptionProps) => {
       noFeedback={selected}
       key={touchableProps.key}
       styles={{
-        feedback: variantStyles.buttonFeedback,
+        // feedback: variantStyles.buttonFeedback,
       }}
-      style={[variantStyles.button, selected && variantStyles['button:selected'], style]}
+      // style={[variantStyles.button, selected && variantStyles['button:selected'], style]}
       onPress={onPress}
     >
       <Text
         text={label}
         style={[
-          variantStyles.text,
-          selected && variantStyles['text:selected'],
-          touchableProps?.disabled && variantStyles['text:disabled'],
+          // variantStyles.text,
+          // selected && variantStyles['text:selected'],
+          // touchableProps?.disabled && variantStyles['text:disabled'],
         ]}
         {...textProps}
       />
