+++ packages/web/src/components/SegmentedControl/index.tsx
@@ -7,16 +7,51 @@ import {
   TypeGuards,
   useDefaultComponentStyle,
 } from '@codeleap/common'
-import { SegmentedControlProps } from './types'
+// import { SegmentedControlProps } from './types'
+// import { SegmentedControlPresets } from './styles'
+
+export * from './styles'
+
+import { PropsOf } from '@codeleap/common'
+// import { StylesOf } from '../../types'
+import { ActionIcon } from '../ActionIcon'
+import { SegmentedControlPresets } from './styles'
+
+export type SegmentedControlProps = {
+  options?: {label: string; value: any }[]
+  /**
+   * Determine if the modal is visible
+  */
+  value?: any
+  /**
+   * Determine if the modal is visible
+  */
+  styles?: React.ReactNode
+  /**
+    * Determine if the modal is visible
+  */
+  variants?: ComponentVariants<typeof SegmentedControlPresets>['variants']
+
+  label?: React.ReactNode
+  /**
+    * Determine if the modal is visible
+  */
+  getItemWidth?: (item:{label: string; value: any }, idx: number, arr: {label: string; value: any }[]) => number
+  /**
+    * Determine if the modal is visible
+  */
+}
 
 export const SegmentedControl = (props: SegmentedControlProps) => {
 
-  const { getItemWidth, label, options, styles, value } = props
+  const { getItemWidth, label, options, styles, value, variants, responsiveVariants } = props
 
-  const variantStyles = useDefaultComponentStyle<'u:SegmentedControl', typeof RadioInputPresets>('u:SegmentedControl', {
-    variants,
-    styles,
-  })
+  const variantStyles = useDefaultComponentStyle<'u:SegmentedControl', typeof SegmentedControlPresets>(
+    'u:SegmentedControl',
+    {
+      variants, styles, responsiveVariants, rootElement: 'wrapper',
+    },
+  )
 
   return (
     <View>
@@ -28,7 +63,7 @@ export const SegmentedControl = (props: SegmentedControlProps) => {
           key={idx}
           // style={widthStyle}
           selected={value === o.value}
-          variantStyles={variantStyles}
+          // variantStyles={variantStyles}
         />
       ))}
     </View>
