+++ packages/web/src/components/SegmentedControl/styles.ts
@@ -1,6 +1,26 @@
 import { createDefaultVariantFactory, includePresets, StylesOf } from '@codeleap/common'
-import { InputLabelComposition } from '../InputLabel'
-import { SegmentedControlComposition, SegmentedControlStylesGen } from './types'
+
+export type SegmentedControlStates = 'selected' | 'disabled'
+
+export type SegmentedControlComposition =
+ 'selectedBubble' |
+ 'wrapper' |
+ 'innerWrapper' |
+ 'scroll' |
+ 'scrollContent' |
+ 'text' |
+ `text:${SegmentedControlStates}` |
+ 'icon' |
+ 'button' |
+ 'buttonFeedback' |
+ `button:${SegmentedControlStates}` |
+  `selectedBubble:${SegmentedControlStates}` |
+ `label`
+
+export type SegmentedControlStylesGen<TCSS = any> =
+  StylesOf<
+    Exclude<SegmentedControlComposition, 'buttonFeedback'>
+  >
 
 const createSegmentedControlStyle = createDefaultVariantFactory<
 SegmentedControlComposition,
@@ -8,3 +28,4 @@ SegmentedControlStylesGen
 >()
 
 export const SegmentedControlPresets = includePresets((style) => createSegmentedControlStyle(() => ({ wrapper: style })))
+
