+++ packages/web/src/components/defaultStyles.ts
@@ -22,6 +22,7 @@ import { LoadingOverlayPresets } from './LoadingOverlay/styles'
 import { InputBasePresets } from './InputBase'
 import { SwitchPresets } from './Switch/styles'
 import { NumberIncrementPresets } from './NumberIncrement/styles'
+import { SegmentedControlPresets } from './SegmentedControl/styles'
 
 export const defaultStyles = {
   View: ViewPresets,
@@ -47,5 +48,6 @@ export const defaultStyles = {
   LoadingOverlay: LoadingOverlayPresets,
   InputBase: InputBasePresets,
   Switch: SwitchPresets,
-  NumberIncrement: NumberIncrementPresets
+  NumberIncrement: NumberIncrementPresets,
+  SegmentedControlPresets: SegmentedControlPresets,
 }
